refactor: Move some YJS test helper functions to their own file
